requirements.txt
python-telegram-bot==21.9
python-dotenv
